refactor(mcp): extract explain tool module
