+++ openspec/changes/refactor-mcp-explain-tool-module/README.md
@@ -1,0 +1,3 @@
+# refactor-mcp-explain-tool-module
+
+Refactor: extract MCP explain tool implementation into a dedicated module
